Handle side-effect imports
Earlier, side effect imports like import 'abc' was neglected because of a pre-emptive check being done too early (which reqd the import clause to be present).
It is not needed to mark the expression as restricted, but only needed for later steps.
So, the consition is moved between these two steps now.